Implement searchable product page with bottom-aligned filter panel and range slider
