+++ mobile/AbdrehimMisbah/ecommerce_ui_replication/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'product_data.dart'; // <-- import your product list and class
 import 'details.dart';
 import 'addUpdate.dart';
+import 'reusable.dart';
+import 'search.dart';
 
 void main() {
   runApp(MyApp());
@@ -12,10 +14,7 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: AddUpdatePage(),
-    );
+    return MaterialApp(debugShowCheckedModeBanner: false, home: SearchPage());
   }
 }
 
@@ -53,120 +52,6 @@ class HomePage extends StatelessWidget {
         ],
       ),
     );
-  }
-
-  List<Widget> productCard() {
-    List<Widget> headingAndProductCard = [];
-
-    // Add heading widget
-    // headingAndProductCard.add(productHeading());
-
-    // Add product cards
-    headingAndProductCard.addAll(
-      products.map((product) {
-        return Card(
-          color: Colors.white,
-          clipBehavior: Clip.antiAlias,
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisAlignment: MainAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              // Product Image
-              AspectRatio(
-                aspectRatio: 18 / 11,
-                child: Image.asset(product.image, fit: BoxFit.fitWidth),
-              ),
-              const SizedBox(height: 8),
-
-              // Name + Price
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 12.0,
-                  vertical: 4.0,
-                ),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text(
-                      product.name,
-                      style: const TextStyle(
-                        fontFamily: 'Poppins', // Custom font family
-                        fontSize: 20,
-                        fontWeight: FontWeight.w500, // Medium weight
-                        color: Color(0xFF3E3E3E), // Hex color #3E3E3E
-                      ),
-                    ),
-
-                    Text(
-                      '\$${product.price.toStringAsFixed(2)}',
-                      style: const TextStyle(
-                        fontFamily: 'Poppins', // Custom font family
-                        fontSize: 14,
-                        fontWeight: FontWeight.w500, // Medium weight
-                        color: Color(0xFF3E3E3E),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-
-              // Type + Rating
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 12.0,
-                  vertical: 4.0,
-                ),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text(
-                      product.type,
-                      style: const TextStyle(
-                        fontSize: 12,
-                        fontFamily: 'Poppins', // Use Poppins font
-                        fontWeight: FontWeight.w400, // Regular weight
-                        color: Color(0xFFAAAAAA), // Hex color #AAAAAA
-                      ),
-                    ),
-
-                    Text.rich(
-                      TextSpan(
-                        children: [
-                          WidgetSpan(
-                            alignment: PlaceholderAlignment.middle,
-                            child: Icon(
-                              Icons.star,
-                              color: Color(0xFFFFD700),
-                              size: 20,
-                            ),
-                          ),
-                          WidgetSpan(
-                            child: SizedBox(width: 4), // spacing
-                          ),
-                          TextSpan(
-                            text: '(${product.rating})',
-                            style: const TextStyle(
-                              fontSize: 12,
-                              fontFamily: 'Sora', // Use Sora font
-                              fontWeight: FontWeight.w400, // Regular weight
-                              color: Color(0xFFAAAAAA), // Hex color #AAAAAA
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 8),
-            ],
-          ),
-        );
-      }).toList(), // 👈 make sure this ends without a comma
-    );
-
-    return headingAndProductCard;
   }
 
   Widget userIntro() {
